requirements.txt
aiogram
pillow
pillow-heif
